Take even more care in converting uintN_t to intN_t
This code converts positive and negative values from uintN_t to intN_t
separately, but was being undone by having different types on the two
sides of the colon, and hence their common type being uintN_t -- thus
requiring an implicit conversion to intN_t after all!

(int8_t and int16_t are smaller than int so this matters even less
for these types, but it's good to follow the pattern anyway.)

Uncovered via -Wsign-compare, and grepping the code for /:.*(int/ etc.
Followup to dc3303b6da542f74c965b3f3f50a58776f87104b.

+++ cram/cram_io.h
@@ -293,7 +293,7 @@ static inline int safe_itf8_get(const char *cp, const char *endp,
         return 4;
     } else {
         uint32_t uv = (((uint32_t)up[0] & 0x0f)<<28) | (up[1]<<20) | (up[2]<<12) | (up[3]<<4) | (up[4] & 0x0f);
-        *val_p = uv < 0x80000000UL ? uv : -((int32_t) (0xffffffffUL - uv)) - 1;
+        *val_p = uv < 0x80000000UL ? (int32_t) uv : -((int32_t) (0xffffffffUL - uv)) - 1;
         return 5;
     }
 }
